requirements.txt
absl-py @ file:///croot/absl-py_1714140470852/work
aiohttp @ file:///croot/aiohttp_1715108769648/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
attrs @ file:///work/perseverance-python-buildout/croot/attrs_1698845858439/work
blinker @ file:///work/perseverance-python-buildout/croot/blinker_1698846408677/work
Bottleneck @ file:///croot/bottleneck_1709069899917/work
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools @ file:///croot/cachetools_1713977085444/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///work/perseverance-python-buildout/croot/click_1698845879718/work
cryptography @ file:///croot/cryptography_1714660666131/work
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
frozenlist @ file:///work/perseverance-python-buildout/croot/frozenlist_1698886842990/work
google-auth @ file:///croot/google-auth_1715111149956/work
google-auth-oauthlib==0.4.1
grpcio @ file:///croot/grpc-suite_1701982571719/work
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///work/perseverance-python-buildout/croot/jinja2_1707343043683/work
Markdown @ file:///work/perseverance-python-buildout/croot/markdown_1698870813381/work
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
multidict @ file:///croot/multidict_1701096859099/work
networkx @ file:///work/perseverance-python-buildout/croot/networkx_1698865062738/work
numexpr @ file:///work/perseverance-python-buildout/croot/numexpr_1698871031164/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
oauthlib @ file:///work/perseverance-python-buildout/croot/oauthlib_1698865093348/work
pandas @ file:///croot/pandas_1709590491089/work/dist/pandas-2.2.1-cp312-cp312-linux_x86_64.whl#sha256=5802047a01a8af8f4a2d6aa0e3e9e6a659b7dcb47f027e62c94eab8c10d85ba7
pillow @ file:///croot/pillow_1714398848491/work
protobuf==3.20.3
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///croot/pyjwt_1715094751094/work
pyOpenSSL @ file:///croot/pyopenssl_1708386183891/work
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///croot/scipy_1714069776893/work/dist/scipy-1.13.0-cp312-cp312-linux_x86_64.whl#sha256=36174fb5b942e39e4e12106540ad23beec20867e0991f2f0a890e68dcfdd80b4
setuptools==69.5.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy @ file:///croot/sympy_1701397643339/work
tensorboard @ file:///home/builder/adipietro/tf/tensorboard_1630508887863/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
torch==2.3.0
torchvision==0.18.0
tqdm @ file:///croot/tqdm_1716395931952/work
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1715635851070/work
Werkzeug @ file:///croot/werkzeug_1716533313107/work
wheel==0.43.0
yarl @ file:///croot/yarl_1701105127787/work
